+++ clicks/usbtoserial/lib_usbtoserial/include/usbtoserial.h
@@ -1,0 +1,243 @@
+/****************************************************************************
+** Copyright (C) 2026 MikroElektronika d.o.o.
+** Contact: https://www.mikroe.com/contact
+**
+** Permission is hereby granted, free of charge, to any person obtaining a copy
+** of this software and associated documentation files (the "Software"), to deal
+** in the Software without restriction, including without limitation the rights
+** to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+** copies of the Software, and to permit persons to whom the Software is
+** furnished to do so, subject to the following conditions:
+** The above copyright notice and this permission notice shall be
+** included in all copies or substantial portions of the Software.
+**
+** THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND,
+** EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES
+** OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT.
+** IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM,
+** DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT
+** OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE
+**  USE OR OTHER DEALINGS IN THE SOFTWARE.
+****************************************************************************/
+
+/*!
+ * @file usbtoserial.h
+ * @brief This file contains API for USB to Serial Click Driver.
+ */
+
+#ifndef USBTOSERIAL_H
+#define USBTOSERIAL_H
+
+#ifdef __cplusplus
+extern "C"{
+#endif
+
+/**
+ * Any initialization code needed for MCU to function properly.
+ * Do not remove this line or clock might not be set correctly.
+ */
+#ifdef PREINIT_SUPPORTED
+#include "preinit.h"
+#endif
+
+#ifdef MikroCCoreVersion
+    #if MikroCCoreVersion >= 1
+        #include "delays.h"
+    #endif
+#endif
+
+#include "drv_digital_out.h"
+#include "drv_digital_in.h"
+#include "drv_uart.h"
+
+/*!
+ * @addtogroup usbtoserial USB to Serial Click Driver
+ * @brief API for configuring and manipulating USB to Serial Click driver.
+ * @{
+ */
+
+/**
+ * @defgroup usbtoserial_cmd USB to Serial Device Settings
+ * @brief Settings of USB to Serial Click driver.
+ */
+
+/**
+ * @addtogroup usbtoserial_cmd
+ * @{
+ */
+
+/**
+ * @brief USB to Serial driver buffer size.
+ * @details Specified size of driver ring buffer.
+ * @note Increase buffer size if needed.
+ */
+#define USBTOSERIAL_TX_DRV_BUFFER_SIZE    200
+#define USBTOSERIAL_RX_DRV_BUFFER_SIZE    200
+
+/*! @} */ // usbtoserial_cmd
+
+/**
+ * @defgroup usbtoserial_map USB to Serial MikroBUS Map
+ * @brief MikroBUS pin mapping of USB to Serial Click driver.
+ */
+
+/**
+ * @addtogroup usbtoserial_map
+ * @{
+ */
+
+/**
+ * @brief MikroBUS pin mapping.
+ * @details Mapping pins of USB to Serial Click to the selected MikroBUS.
+ */
+#define USBTOSERIAL_MAP_MIKROBUS( cfg, mikrobus ) \
+    cfg.tx_pin = MIKROBUS( mikrobus, MIKROBUS_TX ); \
+    cfg.rx_pin = MIKROBUS( mikrobus, MIKROBUS_RX ); \
+    cfg.rst = MIKROBUS( mikrobus, MIKROBUS_RST ); \
+    cfg.wkp = MIKROBUS( mikrobus, MIKROBUS_PWM );
+
+/*! @} */ // usbtoserial_map
+/*! @} */ // usbtoserial
+
+/**
+ * @brief USB to Serial Click context object.
+ * @details Context object definition of USB to Serial Click driver.
+ */
+typedef struct
+{
+    // Output pins
+    digital_out_t rst;                          /**< Reset pin(active low). */
+    digital_out_t wkp;                          /**< Wake-up pin. */
+
+    // Modules
+    uart_t uart;                                /**< UART driver object. */
+
+    // Buffers
+    uint8_t uart_rx_buffer[ USBTOSERIAL_RX_DRV_BUFFER_SIZE ];  /**< RX Buffer size. */
+    uint8_t uart_tx_buffer[ USBTOSERIAL_TX_DRV_BUFFER_SIZE ];  /**< TX Buffer size. */
+
+} usbtoserial_t;
+
+/**
+ * @brief USB to Serial Click configuration object.
+ * @details Configuration object definition of USB to Serial Click driver.
+ */
+typedef struct
+{
+    // Communication gpio pins
+    pin_name_t rx_pin;                          /**< RX pin. */
+    pin_name_t tx_pin;                          /**< TX pin. */
+
+    // Additional gpio pins
+    pin_name_t rst;                             /**< Reset pin descriptor. */
+    pin_name_t wkp;                             /**< Wake-up pin descriptor. */
+
+    // Static variable
+    uint32_t         baud_rate;                 /**< Clock speed. */
+    bool             uart_blocking;             /**< Wait for interrupt or not. */
+    uart_data_bits_t data_bit;                  /**< Data bits. */
+    uart_parity_t    parity_bit;                /**< Parity bit. */
+    uart_stop_bits_t stop_bit;                  /**< Stop bits. */
+
+} usbtoserial_cfg_t;
+
+/**
+ * @brief USB to Serial Click return value data.
+ * @details Predefined enum values for driver return values.
+ */
+typedef enum
+{
+    USBTOSERIAL_OK = 0,
+    USBTOSERIAL_ERROR = -1
+
+} usbtoserial_return_value_t;
+
+/*!
+ * @addtogroup usbtoserial USB to Serial Click Driver
+ * @brief API for configuring and manipulating USB to Serial Click driver.
+ * @{
+ */
+
+/**
+ * @brief USB to Serial configuration object setup function.
+ * @details This function initializes Click configuration structure to initial
+ * values.
+ * @param[out] cfg : Click configuration structure.
+ * See #usbtoserial_cfg_t object definition for detailed explanation.
+ * @return Nothing.
+ * @note The all used pins will be set to unconnected state.
+ */
+void usbtoserial_cfg_setup ( usbtoserial_cfg_t *cfg );
+
+/**
+ * @brief USB to Serial initialization function.
+ * @details This function initializes all necessary pins and peripherals used
+ * for this Click board.
+ * @param[out] ctx : Click context object.
+ * See #usbtoserial_t object definition for detailed explanation.
+ * @param[in] cfg : Click configuration structure.
+ * See #usbtoserial_cfg_t object definition for detailed explanation.
+ * @return @li @c  0 - Success,
+ *         @li @c -1 - Error.
+ * See #err_t definition for detailed explanation.
+ * @note None.
+ */
+err_t usbtoserial_init ( usbtoserial_t *ctx, usbtoserial_cfg_t *cfg );
+
+/**
+ * @brief USB to Serial data writing function.
+ * @details This function writes a desired number of data bytes by using UART serial interface.
+ * @param[in] ctx : Click context object.
+ * See #usbtoserial_t object definition for detailed explanation.
+ * @param[in] data_in : Data buffer for sending.
+ * @param[in] len : Number of bytes for sending.
+ * @return @li @c  >=0 - Success,
+ *         @li @c   <0 - Error.
+ * See #err_t definition for detailed explanation.
+ * @note None.
+ */
+err_t usbtoserial_generic_write ( usbtoserial_t *ctx, uint8_t *data_in, uint16_t len );
+
+/**
+ * @brief USB to Serial data reading function.
+ * @details This function reads a desired number of data bytes by using UART serial interface.
+ * @param[in] ctx : Click context object.
+ * See #usbtoserial_t object definition for detailed explanation.
+ * @param[out] data_out : Output read data.
+ * @param[in] len : Number of bytes to be read.
+ * @return @li @c  >0 - Number of data bytes read,
+ *         @li @c <=0 - Error/Empty Ring buffer.
+ * See #err_t definition for detailed explanation.
+ * @note None.
+ */
+err_t usbtoserial_generic_read ( usbtoserial_t *ctx, uint8_t *data_out, uint16_t len );
+
+/**
+ * @brief USB to Serial reset device function.
+ * @details This function resets the device by toggling the reset pin logic state.
+ * @param[in] ctx : Click context object.
+ * See #usbtoserial_t object definition for detailed explanation.
+ * @return None.
+ * @note None.
+ */
+void usbtoserial_reset_device ( usbtoserial_t *ctx );
+
+/**
+ * @brief USB to Serial set WKP pin function.
+ * @details This function sets the WKP pin to the selected logic state.
+ * @param[in] ctx : Click context object.
+ * See #usbtoserial_t object definition for detailed explanation.
+ * @param[in] state : Pin logic state.
+ * @return None.
+ * @note None.
+ */
+void usbtoserial_set_wkp_pin ( usbtoserial_t *ctx, uint8_t state );
+
+#ifdef __cplusplus
+}
+#endif
+#endif // USBTOSERIAL_H
+
+/*! @} */ // usbtoserial
+
+// ------------------------------------------------------------------------ END